apollo_starknet_client: inline StarknetClient into StarknetFeederGatewayClient
Move fields (http_headers, internal_client, retry_config) and methods
(request_with_retry, request, get_retry_error_code, should_retry) from the
now-deleted StarknetClient directly onto StarknetFeederGatewayClient.
Relocate request_with_retry tests into reader/request_with_retry_test.rs.
