+++ crates/apollo_starknet_client/src/reader/request_with_retry_test.rs
@@ -3,7 +3,9 @@ use reqwest::StatusCode;
 
 use crate::starknet_error::{KnownStarknetErrorCode, StarknetError, StarknetErrorCode};
 use crate::test_utils::retry::{get_test_config, MAX_RETRIES};
-use crate::{ClientError, RetryErrorCode, StarknetClient};
+use crate::{ClientError, RetryErrorCode};
+
+use super::StarknetFeederGatewayClient;
 
 const NODE_VERSION: &str = "NODE VERSION";
 const URL_SUFFIX: &str = "/query";
@@ -11,14 +13,13 @@ const URL_SUFFIX: &str = "/query";
 #[tokio::test]
 async fn request_with_retry_positive_flow() {
     const BODY: &str = "body";
-    let apollo_starknet_client =
-        StarknetClient::new(None, NODE_VERSION, get_test_config()).unwrap();
     let mut server = mockito::Server::new_async().await;
+    let client =
+        StarknetFeederGatewayClient::new(&server.url(), None, NODE_VERSION, get_test_config())
+            .unwrap();
     let mock = server.mock("GET", URL_SUFFIX).with_status(200).with_body(BODY).create_async().await;
     let url = format!("{}{}", server.url(), URL_SUFFIX);
-    let result = apollo_starknet_client
-        .request_with_retry(apollo_starknet_client.internal_client.get(&url))
-        .await;
+    let result = client.request_with_retry(client.internal_client.get(&url)).await;
     assert_eq!(result.unwrap(), BODY);
     mock.assert_async().await;
 }
@@ -26,15 +27,14 @@ async fn request_with_retry_positive_flow() {
 #[tokio::test]
 async fn request_with_retry_bad_response_status() {
     let error_code = StatusCode::NOT_FOUND;
-    let apollo_starknet_client =
-        StarknetClient::new(None, NODE_VERSION, get_test_config()).unwrap();
     let mut server = mockito::Server::new_async().await;
+    let client =
+        StarknetFeederGatewayClient::new(&server.url(), None, NODE_VERSION, get_test_config())
+            .unwrap();
     let mock =
         server.mock("GET", URL_SUFFIX).with_status(error_code.as_u16().into()).create_async().await;
     let url = format!("{}{}", server.url(), URL_SUFFIX);
-    let result = apollo_starknet_client
-        .request_with_retry(apollo_starknet_client.internal_client.get(&url))
-        .await;
+    let result = client.request_with_retry(client.internal_client.get(&url)).await;
     assert_matches!(
         result,
         Err(ClientError::BadResponseStatus { code, message: _ }) if code == error_code
@@ -44,13 +44,14 @@ async fn request_with_retry_bad_response_status() {
 
 #[tokio::test]
 async fn request_with_retry_starknet_error_no_retry() {
-    let apollo_starknet_client =
-        StarknetClient::new(None, NODE_VERSION, get_test_config()).unwrap();
+    let mut server = mockito::Server::new_async().await;
+    let client =
+        StarknetFeederGatewayClient::new(&server.url(), None, NODE_VERSION, get_test_config())
+            .unwrap();
     let expected_starknet_error = StarknetError {
         code: StarknetErrorCode::KnownErrorCode(KnownStarknetErrorCode::UndeclaredClass),
         message: "message".to_string(),
     };
-    let mut server = mockito::Server::new_async().await;
     let mock = server
         .mock("GET", URL_SUFFIX)
         .with_status(StatusCode::BAD_REQUEST.as_u16().into())
@@ -58,9 +59,7 @@ async fn request_with_retry_starknet_error_no_retry() {
         .create_async()
         .await;
     let url = format!("{}{}", server.url(), URL_SUFFIX);
-    let result = apollo_starknet_client
-        .request_with_retry(apollo_starknet_client.internal_client.get(&url))
-        .await;
+    let result = client.request_with_retry(client.internal_client.get(&url)).await;
     let Err(ClientError::StarknetError(starknet_error)) = result else {
         panic!("Did not get a StarknetError.");
     };
@@ -70,9 +69,10 @@ async fn request_with_retry_starknet_error_no_retry() {
 
 #[tokio::test]
 async fn request_with_retry_serde_error_in_starknet_error() {
-    let apollo_starknet_client =
-        StarknetClient::new(None, NODE_VERSION, get_test_config()).unwrap();
     let mut server = mockito::Server::new_async().await;
+    let client =
+        StarknetFeederGatewayClient::new(&server.url(), None, NODE_VERSION, get_test_config())
+            .unwrap();
     let mock = server
         .mock("GET", URL_SUFFIX)
         .with_status(StatusCode::BAD_REQUEST.as_u16().into())
@@ -80,17 +80,17 @@ async fn request_with_retry_serde_error_in_starknet_error() {
         .create_async()
         .await;
     let url = format!("{}{}", server.url(), URL_SUFFIX);
-    let result = apollo_starknet_client
-        .request_with_retry(apollo_starknet_client.internal_client.get(&url))
-        .await;
+    let result = client.request_with_retry(client.internal_client.get(&url)).await;
     assert_matches!(result, Err(ClientError::SerdeError(_)));
     mock.assert_async().await;
 }
 
 #[tokio::test]
 async fn request_with_retry_max_retries_reached() {
-    let apollo_starknet_client =
-        StarknetClient::new(None, NODE_VERSION, get_test_config()).unwrap();
+    let mut server = mockito::Server::new_async().await;
+    let client =
+        StarknetFeederGatewayClient::new(&server.url(), None, NODE_VERSION, get_test_config())
+            .unwrap();
     for (status_code, error_code) in [
         (StatusCode::TEMPORARY_REDIRECT, RetryErrorCode::Redirect),
         (StatusCode::REQUEST_TIMEOUT, RetryErrorCode::Timeout),
@@ -98,7 +98,6 @@ async fn request_with_retry_max_retries_reached() {
         (StatusCode::SERVICE_UNAVAILABLE, RetryErrorCode::ServiceUnavailable),
         (StatusCode::GATEWAY_TIMEOUT, RetryErrorCode::Timeout),
     ] {
-        let mut server = mockito::Server::new_async().await;
         let mock = server
             .mock("GET", URL_SUFFIX)
             .with_status(status_code.as_u16().into())
@@ -106,9 +105,7 @@ async fn request_with_retry_max_retries_reached() {
             .create_async()
             .await;
         let url = format!("{}{}", server.url(), URL_SUFFIX);
-        let result = apollo_starknet_client
-            .request_with_retry(apollo_starknet_client.internal_client.get(&url))
-            .await;
+        let result = client.request_with_retry(client.internal_client.get(&url)).await;
         assert_matches!(
             result, Err(ClientError::RetryError { code, message: _ }) if code == error_code
         );
@@ -120,8 +117,10 @@ async fn request_with_retry_max_retries_reached() {
 async fn request_with_retry_success_on_retry() {
     const BODY: &str = "body";
     assert_ne!(0, MAX_RETRIES);
-    let apollo_starknet_client =
-        StarknetClient::new(None, NODE_VERSION, get_test_config()).unwrap();
+    let mut server = mockito::Server::new_async().await;
+    let client =
+        StarknetFeederGatewayClient::new(&server.url(), None, NODE_VERSION, get_test_config())
+            .unwrap();
     for status_code in [
         StatusCode::TEMPORARY_REDIRECT,
         StatusCode::REQUEST_TIMEOUT,
@@ -129,7 +128,6 @@ async fn request_with_retry_success_on_retry() {
         StatusCode::SERVICE_UNAVAILABLE,
         StatusCode::GATEWAY_TIMEOUT,
     ] {
-        let mut server = mockito::Server::new_async().await;
         let mock_failure = server
             .mock("GET", URL_SUFFIX)
             .with_status(status_code.as_u16().into())
@@ -139,9 +137,7 @@ async fn request_with_retry_success_on_retry() {
         let mock_success =
             server.mock("GET", URL_SUFFIX).with_status(200).with_body(BODY).create_async().await;
         let url = format!("{}{}", server.url(), URL_SUFFIX);
-        let result = apollo_starknet_client
-            .request_with_retry(apollo_starknet_client.internal_client.get(&url))
-            .await;
+        let result = client.request_with_retry(client.internal_client.get(&url)).await;
         assert_eq!(result.unwrap(), BODY);
         mock_failure.assert_async().await;
         mock_success.assert_async().await;
@@ -150,14 +146,15 @@ async fn request_with_retry_success_on_retry() {
 
 #[tokio::test]
 async fn request_with_retry_starknet_error_max_retries_reached() {
-    let apollo_starknet_client =
-        StarknetClient::new(None, NODE_VERSION, get_test_config()).unwrap();
+    let mut server = mockito::Server::new_async().await;
+    let client =
+        StarknetFeederGatewayClient::new(&server.url(), None, NODE_VERSION, get_test_config())
+            .unwrap();
     let starknet_error = StarknetError {
         code: StarknetErrorCode::KnownErrorCode(KnownStarknetErrorCode::TransactionLimitExceeded),
         message: "message".to_string(),
     };
     let starknet_error_str = serde_json::to_string(&starknet_error).unwrap();
-    let mut server = mockito::Server::new_async().await;
     let mock = server
         .mock("GET", URL_SUFFIX)
         .with_status(StatusCode::BAD_REQUEST.as_u16().into())
@@ -166,9 +163,7 @@ async fn request_with_retry_starknet_error_max_retries_reached() {
         .create_async()
         .await;
     let url = format!("{}{}", server.url(), URL_SUFFIX);
-    let result = apollo_starknet_client
-        .request_with_retry(apollo_starknet_client.internal_client.get(&url))
-        .await;
+    let result = client.request_with_retry(client.internal_client.get(&url)).await;
     assert_matches!(
         result,
         Err(ClientError::RetryError { code, message: _ }) if code == RetryErrorCode::TooManyRequests
@@ -180,14 +175,15 @@ async fn request_with_retry_starknet_error_max_retries_reached() {
 async fn request_with_retry_starknet_error_success_on_retry() {
     const BODY: &str = "body";
     assert_ne!(0, MAX_RETRIES);
-    let apollo_starknet_client =
-        StarknetClient::new(None, NODE_VERSION, get_test_config()).unwrap();
+    let mut server = mockito::Server::new_async().await;
+    let client =
+        StarknetFeederGatewayClient::new(&server.url(), None, NODE_VERSION, get_test_config())
+            .unwrap();
     let starknet_error = StarknetError {
         code: StarknetErrorCode::KnownErrorCode(KnownStarknetErrorCode::TransactionLimitExceeded),
         message: "message".to_string(),
     };
     let starknet_error_str = serde_json::to_string(&starknet_error).unwrap();
-    let mut server = mockito::Server::new_async().await;
     let mock_failure = server
         .mock("GET", URL_SUFFIX)
         .with_status(StatusCode::BAD_REQUEST.as_u16().into())
@@ -198,9 +194,7 @@ async fn request_with_retry_starknet_error_success_on_retry() {
     let mock_success =
         server.mock("GET", URL_SUFFIX).with_status(200).with_body(BODY).create_async().await;
     let url = format!("{}{}", server.url(), URL_SUFFIX);
-    let result = apollo_starknet_client
-        .request_with_retry(apollo_starknet_client.internal_client.get(&url))
-        .await;
+    let result = client.request_with_retry(client.internal_client.get(&url)).await;
     assert_eq!(result.unwrap(), BODY);
     mock_failure.assert_async().await;
     mock_success.assert_async().await;
